feat(infakt): post-create webhook prompt in the connect wizard (#1770)
On a successful inFakt connection create, a modal asks whether to set up
webhooks now; "Configure webhooks" opens the full config dialog, "Not now"
dismisses with a toast that names where to find it later (Actions tab). The
post-create region keeps a "Configure webhooks..." button to reopen it.

Moves the InfaktWebhookConfig body + dialog from the inFakt plugin into the
connections feature so the wizard can render it (a feature can't import a
plugin per the ESLint boundary; a plugin can import the feature). The
plugin's ConnectionActions slot now wraps the feature-owned dialog. Tests
split accordingly; existing post-create wizard tests dismiss the prompt.
